stage2_streamlit/requirements.txt
# Web Framework
streamlit>=1.30.0

# YAML Editor
streamlit-ace>=0.1.1

# Visualization
plotly>=5.18.0
networkx>=3.2.0

# Data Processing
pandas>=2.1.0

# Configuration
pyyaml>=6.0

# Data Models
pydantic>=2.5.0

# LLM Integration
anthropic>=0.18.0

# Test Data Generation
faker>=22.0.0

# Additional utilities
python-dotenv>=1.0.0
